refactor: private types module
