+++ src/vallenae/io/__init__.py
@@ -99,6 +99,6 @@ from .datatypes import *
 from .pridb import *
 from .tradb import *
 from .trfdb import *
-from .types import *
+from ._types import *
 
 __all__ = [_ for _ in dir() if not _.startswith("_")]
